api/requirements.txt
deepdiff==8.2.0
fastapi==0.115.11
psycopg2_binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
spotipy==2.25.0
uvicorn==0.34.0
